2.8.7 bug fix
git-svn-id: https://plugins.svn.wordpress.org/rsvpmaker/trunk@825857 b8457f37-d9ea-0310-8a92-e5e31aec5664

+++ rsvpmaker-plugabble.php
@@ -1229,7 +1229,7 @@ $per = unserialize($custom_fields["_per"][0]);
 
 foreach($per["unit"] as $index => $value)
 	{
-	$price = (int) $per["price"][$index];
+	$price = (float) $per["price"][$index];
 	if(!$price)
 		break;
 	$pf .= '<div><select name="payingfor['.$index.']" class="tickets"><option value="0">0</option><option value="1">1</option><option value="2">2</option><option value="3">3</option><option value="4">4</option><option value="5">5</option><option value="6">6</option><option value="7">7</option><option value="8">8</option><option value="9">9</option><option value="10">10</option></select><input type="hidden" name="unit['.$index.']" value="'.$value.'" />'.$value.' @ <input type="hidden" name="price['.$index.']" value="'.$price.'" />'.(($rsvp_options["paypal_currency"] == 'USD') ? '$' : $rsvp_options["paypal_currency"]).' '.number_format($price,2,$rsvp_options["currency_decimal"],$rsvp_options["currency_thousands"]).'</div>'."\n";
